core: Get rid of type mismatch warning
change cur in ofi_bitmask_get_lsbset to size_t to avoid type mismatch warning when comparing mask.size

+++ include/ofi_bitmask.h
@@ -90,7 +90,7 @@ static inline void ofi_bitmask_set_all(struct bitmask *mask)
 
 static inline size_t ofi_bitmask_get_lsbset(struct bitmask mask)
 {
-	int cur;
+	size_t cur;
 	uint8_t tmp;
 	size_t ret = 0;
